aws_grafana_lease_expire_monitoring/script_db_creation/requirements.txt
boto3
mysql.connector
mysql-connector-python